Added shuffle mode
Added two new menu items for random mode and shuffle mode. Shuffle mode shuffles the album list so that you will not get the same album twice until every other album is played. This will reset on a restart, when switching the mode or when the album list is updated from Spotify.
Removed the interfaces as we currently don't need them. We can add them later again when we possibly add unit tests and dependency injection.