Added playlist generation
You can now generate a playlist of all albums in your library.
The maximum number of tracks is set to 9500 for now. If the number of tracks in the albums exceeds this we create a new playlist and repeat this as long as there are more tracks in the list than fit in a playlist.
The created playlists are updated on every run, so you can put them in a folder and then just play the folder if you want to play everything.